배포

+++ src/main/java/com/safeking/shop/domain/user/domain/service/RedisService.java
@@ -12,8 +12,11 @@ public class RedisService {
     // application load 시점에 redis를 한번 비운다.
     public void deleteAll(){
         Iterable<RedisMember> all = redisRepository.findAll();
+
         for (RedisMember redisMember : all) {
-            redisRepository.delete(redisMember);
+            if(redisMember != null) {
+                redisRepository.delete(redisMember);
+            }
         }
     }
 }
